♻️ 追加リファクタリングとid, 住所を含ませた
