+++ src/components/Home/MapParts/StoreInfo.vue
@@ -1,14 +1,14 @@
 <template>
   <div>
     <l-marker
-      v-for="(data, i) in shopData"
-      :key="i"
+      v-for="data in shopData"
+      :key="data.id"
       :lat-lng="[data.lat, data.lng]"
     >
       <l-tooltip
         :options="{ permanent: true, interactive: true, direction: 'top' }"
       >
-        <div @click="handleToolTipClick(i)">
+        <div @click="handleToolTipClick(data.id)">
           {{ data.name }} | ☆<br />
           <p v-show="data.showDetail">
             {{ data.genre }}<br />
@@ -24,6 +24,10 @@
 <script>
 import { Icon } from "leaflet"
 import { LMarker, LTooltip } from "vue2-leaflet"
+// 非同期処理のライブラリ
+import axios from "axios"
+// ホットペッパーのレスポンスがXML形式だったので変換するためのライブラリ
+const xmlToJson = require("xml2js")
 
 // マーカーが表示されないことを回避するために公式から
 delete Icon.Default.prototype._getIconUrl
@@ -35,7 +39,7 @@ Icon.Default.mergeOptions({
 
 export default {
   props: {
-    shopData: {
+    center: {
       type: Array,
       require: true,
     },
@@ -44,9 +48,49 @@ export default {
     LMarker,
     LTooltip,
   },
+  data() {
+    return {
+      shopData: [],
+    }
+  },
   methods: {
-    handleToolTipClick(index) {
-      this.shopData[index].showDetail = !this.shopData[index].showDetail
+    handleToolTipClick(id) {
+      for (let i = 0; i < this.shopData.length; i++) {
+        if (id === this.shopData[i].id) {
+          this.shopData[i].showDetail = !this.shopData[i].showDetail
+          return
+        }
+      }
+    },
+  },
+  watch: {
+    center: async function () {
+      const HOTPEPPER_URL = "/hotpepper/gourmet/v1/"
+      const API_KEY = "e9cd52a7ae2c9ea3"
+      // const jenre = "G002" // TODO ジャンルを指定する
+      const range = "5"
+      const count = "100"
+
+      const res = await axios.get(
+        `${HOTPEPPER_URL}/?key=${API_KEY}&lat=${this.center[0]}&lng=${this.center[1]}&range=${range}&count=${count}`
+      )
+      const data = await xmlToJson.parseStringPromise(res.data)
+
+      data.results.shop.forEach((restaurant) => {
+        const info = {
+          id: restaurant.id[0],
+          lat: restaurant.lat[0],
+          lng: restaurant.lng[0],
+          name: restaurant.name[0],
+          genre: restaurant.genre[0].name[0],
+          address: restaurant.address[0],
+          time: restaurant.open[0],
+          url: restaurant.urls[0].pc[0],
+          photo: restaurant.photo[0].mobile[0].l[0],
+          showDetail: false,
+        }
+        this.shopData.push(info)
+      })
     },
   },
 }
